feat: add errors to the AST.
