+++ parser-ng/src/parser/tests/testdata/basic-error-1.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `meta`, `strings` or `condition`, found `{`", span: Span(11..12) }
