+++ parser-ng/src/parser/tests/testdata/basic-error-2.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting STRING, regexp or `{`, found `\"foo`", span: Span(33..37) }
