+++ parser-ng/src/parser/tests/testdata/for-error-1.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting expression, found `of`", span: Span(77..79) }
