+++ parser-ng/src/parser/tests/testdata/hex-patterns-error-2.ast
@@ -1,0 +1,3 @@
+ERRORS:
+- SyntaxError { message: "expecting `[`, BYTE, `(` or `}`, found `0`", span: Span(33..34) }
+- SyntaxError { message: "expecting `[`, BYTE, `(` or `}`, found `a`", span: Span(93..94) }
