+++ parser-ng/src/parser/tests/testdata/meta-error-1.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `meta`, `strings` or `condition`, found `foo`", span: Span(14..17) }
