+++ parser-ng/src/parser/tests/testdata/meta-error-2.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `=`, found `condition`", span: Span(30..39) }
