+++ parser-ng/src/parser/tests/testdata/meta-error-3.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `-`, INTEGER, FLOAT, STRING, `true` or `false`, found `condition`", span: Span(32..41) }
