+++ parser-ng/src/parser/tests/testdata/meta-error-4.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `-`, INTEGER, FLOAT, STRING, `true` or `false`, found `meta`", span: Span(30..34) }
