+++ parser-ng/src/parser/tests/testdata/of-error-1.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "unexpected `at`", span: Span(41..43) }
