+++ parser-ng/src/parser/tests/testdata/pattern-mods-error-1.ast
@@ -1,0 +1,3 @@
+ERRORS:
+- SyntaxError { message: "expecting pattern modifier, pattern identifier or `condition`, found `foo`", span: Span(36..39) }
+- SyntaxError { message: "expecting STRING, found `)`", span: Span(107..108) }
