+++ parser-ng/src/parser/tests/testdata/patterns-error-1.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `[`, BYTE, `(` or `}`, found `condition`", span: Span(37..46) }
