+++ parser-ng/src/parser/tests/testdata/rule-tags-error-2.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting `:` or `{`, found `=`", span: Span(10..11) }
