+++ parser-ng/src/parser/tests/testdata/rule-tags-error-3.ast
@@ -1,0 +1,2 @@
+ERRORS:
+- SyntaxError { message: "expecting identifier, found `=`", span: Span(12..13) }
